+++ __azurite_db_blob_extent__.json
@@ -1,1 +1,1 @@
-{"filename":"c:\\git\\STU-Copilot\\__azurite_db_blob_extent__.json","collections":[{"name":"$EXTENTS_COLLECTION$","data":[{"id":"cf3e0894-de7c-45bf-b440-85d121d91c26","locationId":"Default","path":"cf3e0894-de7c-45bf-b440-85d121d91c26","size":4334,"lastModifiedInMS":1752840703195,"meta":{"revision":41,"created":1752840703195,"version":0,"updated":1752848351972},"$loki":1,"LastModifyInMS":1752848351972},{"id":"d7a34769-9e77-48f1-923c-e35f1d3221fa","locationId":"Default","path":"d7a34769-9e77-48f1-923c-e35f1d3221fa","size":0,"lastModifiedInMS":1752845960924,"meta":{"revision":0,"created":1752845960925,"version":0},"$loki":2}],"idIndex":[1,2],"binaryIndices":{"id":{"name":"id","dirty":false,"values":[0,1]}},"constraints":null,"uniqueNames":[],"transforms":{},"objType":"$EXTENTS_COLLECTION$","dirty":false,"cachedIndex":null,"cachedBinaryIndex":null,"cachedData":null,"adaptiveBinaryIndices":true,"transactional":false,"cloneObjects":false,"cloneMethod":"parse-stringify","asyncListeners":false,"disableMeta":false,"disableChangesApi":true,"disableDeltaChangesApi":true,"autoupdate":false,"serializableIndices":true,"disableFreeze":true,"ttl":null,"maxId":2,"DynamicViews":[],"events":{"insert":[],"update":[],"pre-insert":[],"pre-update":[],"close":[],"flushbuffer":[],"error":[],"delete":[null],"warning":[null]},"changes":[],"dirtyIds":[]}],"databaseVersion":1.5,"engineVersion":1.5,"autosave":true,"autosaveInterval":5000,"autosaveHandle":null,"throttledSaves":true,"options":{"persistenceMethod":"fs","autosave":true,"autosaveInterval":5000,"serializationMethod":"normal","destructureDelimiter":"$<\n"},"persistenceMethod":"fs","persistenceAdapter":null,"verbose":false,"events":{"init":[null],"loaded":[],"flushChanges":[],"close":[],"changes":[],"warning":[]},"ENV":"NODEJS"}
+{"filename":"c:\\git\\STU-Copilot\\__azurite_db_blob_extent__.json","collections":[{"name":"$EXTENTS_COLLECTION$","data":[{"id":"cf3e0894-de7c-45bf-b440-85d121d91c26","locationId":"Default","path":"cf3e0894-de7c-45bf-b440-85d121d91c26","size":4334,"lastModifiedInMS":1752840703195,"meta":{"revision":41,"created":1752840703195,"version":0,"updated":1752848351972},"$loki":1,"LastModifyInMS":1752848351972},{"id":"d7a34769-9e77-48f1-923c-e35f1d3221fa","locationId":"Default","path":"d7a34769-9e77-48f1-923c-e35f1d3221fa","size":0,"lastModifiedInMS":1752845960924,"meta":{"revision":0,"created":1752845960925,"version":0},"$loki":2}],"idIndex":null,"binaryIndices":{"id":{"name":"id","dirty":false,"values":[0,1]}},"constraints":null,"uniqueNames":[],"transforms":{},"objType":"$EXTENTS_COLLECTION$","dirty":false,"cachedIndex":null,"cachedBinaryIndex":null,"cachedData":null,"adaptiveBinaryIndices":true,"transactional":false,"cloneObjects":false,"cloneMethod":"parse-stringify","asyncListeners":false,"disableMeta":false,"disableChangesApi":true,"disableDeltaChangesApi":true,"autoupdate":false,"serializableIndices":true,"disableFreeze":true,"ttl":null,"maxId":2,"DynamicViews":[],"events":{"insert":[],"update":[],"pre-insert":[],"pre-update":[],"close":[],"flushbuffer":[],"error":[],"delete":[null],"warning":[null]},"changes":[],"dirtyIds":[]}],"databaseVersion":1.5,"engineVersion":1.5,"autosave":true,"autosaveInterval":5000,"autosaveHandle":null,"throttledSaves":true,"options":{"persistenceMethod":"fs","autosave":true,"autosaveInterval":5000,"serializationMethod":"normal","destructureDelimiter":"$<\n"},"persistenceMethod":"fs","persistenceAdapter":null,"verbose":false,"events":{"init":[null],"loaded":[],"flushChanges":[],"close":[],"changes":[],"warning":[]},"ENV":"NODEJS"}
