+++ architecture/knowledge_sources.md
@@ -1,9 +1,9 @@
-- GitHub Repositories
-- Microsoft Docs MCP
-- [Azure Pricing API](https://prices.azure.com:443/api/retail/prices)
-- [Tech Community Blogs](https://techcommunity.microsoft.com/Blogs)
-- [DevBogs](https://devblogs.microsoft.com/landing)
-- [Microsoft News](https://news.microsoft.com/source/feed/)
+[x] GitHub Repositories
+[x] Microsoft Docs MCP
+[ ] [Azure Pricing API](https://prices.azure.com:443/api/retail/prices)
+[ ] [Tech Community Blogs](https://techcommunity.microsoft.com/Blogs)
+[ ] [DevBogs](https://devblogs.microsoft.com/landing)
+[ ] [Microsoft News](https://news.microsoft.com/source/feed/)
