+++ src/crawlers/foundry_service.py
@@ -3,14 +3,9 @@ from openai import AzureOpenAI
 from openai.types import CreateEmbeddingResponse
 import json
 import logging
-from dotenv import load_dotenv
-
-# Load environment variables from .env file
-load_dotenv(override=True)
 
 # Configure logging
-# logging.basicConfig(level=logging.WARNING)
-logger = logging.getLogger(__name__)
+logger = logging.getLogger("azure.functions")
 
 
 class FoundryService:
